requirements.txt
fastapi==0.89.1
python_nmap==0.7.1
uvicorn==0.20.0